s.u.
git-svn-id: https://www.technikum29.de/svn/technikum29-www@1384 388e7c26-a98e-49f6-8d48-46b9a15b27c2

+++ physical-computing/Zufallstoene_Bl.5_Aufg.3/Zufallstoene_Bl.5_Aufg.3.ino
@@ -1,7 +1,7 @@
 // Blatt 5, Aufgabe 3 mit Zusatz und Erweiterung:
 // Einstellbare Zahl der abgespielten Toene in einem Zyklus
 
-int Dauer=200;           // Dauer fuer einen Ton in ms, Start mit 200ms
+float Dauer = 200;       // Dauer fuer einen Ton in ms, Start mit 200ms
 int Zahl = 0;            // Zaehler fuer die Zahl der abgespielten Toene
 int maxZahl = 250;       // maximale Zahl der abgespielten Toene
 
@@ -13,8 +13,8 @@ void loop() {
   int frequenz = random(100, 2000);   // Frequenzbereich von 100 bis 2000 Hz
   tone(8, frequenz);
   delay(Dauer);
-  //Dauer=Dauer-0.4;        // alternativ: 
-  Dauer=Dauer*0.996;        // Vorsicht! Kein Komma verwenden!
+  //Dauer=Dauer-0.4;        // Vorsicht! "Dauer" kann negativ werden. Daher besser: 
+  Dauer=Dauer*0.996;        // 0.996: Kein Komma verwenden!
   Zahl++;                   // Zaehler hochzaehlen
   if(Zahl>maxZahl){
     noTone(8);              // Ton abschalten
